requirements.txt
pillow  
djangorestframework
django-filter

#pip install -r requirements.txt